+++ apps/OpenSpace/ext/launcher/include/launcherwindow.h
@@ -34,7 +34,7 @@
 #include <QApplication>
 #include <optional>
 
-namespace openspace::configuration { struct Configuration; }
+namespace openspace { struct Configuration; }
 
 class QComboBox;
 class QLabel;
@@ -54,8 +54,8 @@ public:
      *                   in the tree structure.
      */
     LauncherWindow(bool profileEnabled,
-        const openspace::configuration::Configuration& globalConfig,
-        bool sgctConfigEnabled,  std::string sgctConfigName, QWidget* parent);
+        const openspace::Configuration& globalConfig, bool sgctConfigEnabled,
+        std::string sgctConfigName, QWidget* parent);
 
     /**
       * Returns bool for whether "start OpenSpace" was chosen when this window closed.
