+++ src/engine/globals.cpp
@@ -89,7 +89,7 @@ namespace {
         sizeof(TimeManager) +
         sizeof(VersionChecker) +
         sizeof(WindowDelegate) +
-        sizeof(configuration::Configuration) +
+        sizeof(Configuration) +
         sizeof(interaction::ActionManager) +
         sizeof(interaction::InteractionMonitor) +
         sizeof(interaction::WebsocketInputStates) +
@@ -266,11 +266,11 @@ void create() {
 #endif // WIN32
 
 #ifdef WIN32
-    configuration = new (currentPos) configuration::Configuration;
+    configuration = new (currentPos) Configuration;
     ghoul_assert(configuration, "No configuration");
-    currentPos += sizeof(configuration::Configuration);
+    currentPos += sizeof(Configuration);
 #else // ^^^ WIN32 / !WIN32 vvv
-    configuration = new configuration::Configuration;
+    configuration = new Configuration;
 #endif // WIN32
 
 #ifdef WIN32
